moving modal comp to a specific dir
