+++ src/components/FoodOrder/Cart/Cart.js
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from "react";
-import Modal from "../UI/Modal";
+import Modal from "../UI/Modal/Modal";
 import CartItem from "./CartItem";
 import classes from "./Cart.module.css";
 import CartContext from "../store/cart-context";
