+++ src/components/FoodOrder/Login/Signup.js
@@ -1,5 +1,5 @@
 import React, { useState, useRef } from "react";
-import Modal from "../UI/Modal";
+import Modal from "../UI/Modal/Modal";
 import classes from "./Login.module.css";
 import Input from "../UI/Input/Input";
